refactor: introduce a standalone class for confluence check
